tests: Replace tempdir with tempfile (#380)
The tempdir crate, a dev-dependency needed for the tests, has been
deprecated in favor of tempfile.  This patch replaces tempdir 0.3.7 with
tempfile 3.1.0.

Note that this patch also removes the prefixes from the temporary
directories: They are no longer required by the API, and they are not
needed for the tests.

----

I wrote this patch for the Debian package as the deprecated tempdir
crate will not be packaged for Debian.  But I think it would make sense
to also replace the dependency upstream.